requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2021.5.30
charset-normalizer==2.0.3
click==8.0.1
colorama==0.4.4
Flask==2.0.1
idna==3.2
itsdangerous==2.0.1
Jinja2==3.0.1
lxml==4.6.3
MarkupSafe==2.0.1
mysql-connector-python==8.0.25
protobuf==3.17.3
requests==2.26.0
selenium==3.141.0
six==1.16.0
soupsieve==2.2.1
Unidecode==1.2.0
urllib3==1.26.6
Werkzeug==2.0.1
